optimization: loop counters type made signed int

+++ src/hc/rmd160.cu
@@ -92,7 +92,7 @@ __device__ __forceinline__ BOOL prrmd160_compare(unsigned char* password, const 
     BOOL result = TRUE;
 
 #pragma unroll (4)
-    for (size_t i = 0; i < HASH_LEN && result; ++i) {
+    for (int i = 0; i < HASH_LEN && result; ++i) {
         result &= hash[i] == k_hash[i];
     }
